requirements.txt
alpaca-py
openai
python-dotenv
pandas
numpy
scikit-learn
xgboost
yfinance
beautifulsoup4
requests
notebook
matplotlib
seaborn
vaderSentiment
google-generativeai
openai
anthropic
pytz
pytest>=7.0.0
pytest-cov>=4.0.0
pydantic